Separate out pkg/network/status for shared gateway probing.
This refactors the functionality of the ingress prober to tease apart the probing logic from the logic that resolves the ingress-specific gateway information about what to probe.

The intent of this is to share the core probing logic with one small shim per ingress implementation.
